update: 누적시간 바로 반영 완
