+++ src/app/homepage/_component/D-Day.tsx
@@ -1,37 +1,26 @@
 'use client';
 import React, { useEffect, useState, useRef } from 'react';
 import { supabase } from '@/lib/supabaseClient';
-
-interface DdayItem {
-  id: string;
-  title: string;
-  dday: number;
-  accumulatedTime?: number;
-}
+import { DdayItem } from '../page';
 
 interface DdayListProps {
   onSelectDday: (id: string) => void;
   selectedDdayId: string | null;
-  ddayTimes: { [key: string]: number };
+  ddayList: DdayItem[];
+  calculateDday: (selectedDate: string) => number;
+  setDdayList: (value: React.SetStateAction<DdayItem[]>) => void;
 }
 
 const DdayList: React.FC<DdayListProps> = ({
   onSelectDday,
   selectedDdayId,
-  ddayTimes,
+  ddayList,
+  calculateDday,
+  setDdayList,
 }) => {
   const [formVisible, setFormVisible] = useState(false);
   const [title, setTitle] = useState('');
   const [date, setDate] = useState('');
-  const [ddayList, setDdayList] = useState<DdayItem[]>([]);
-
-  // D-day 계산
-  const calculateDday = (selectedDate: string) => {
-    const today = new Date();
-    const targetDate = new Date(selectedDate);
-    const timeDifference = targetDate.getTime() - today.getTime();
-    return Math.ceil(timeDifference / (1000 * 60 * 60 * 24));
-  };
 
   // Supabase에서 새로운 D-day 항목을 추가하는 함수
   const handleSubmit = async (e: React.FormEvent) => {
@@ -78,13 +67,13 @@ const DdayList: React.FC<DdayListProps> = ({
     setFormVisible(false);
   };
 
-  const formatTime = (time: number) => {
-    const getSeconds = `0${time % 60}`.slice(-2);
-    const minutes = `${Math.floor(time / 60)}`;
-    const getMinutes = `0${Number(minutes) % 60}`.slice(-2);
-    const getHours = `0${Math.floor(time / 3600)}`.slice(-2);
-    return `${getHours}:${getMinutes}:${getSeconds}`;
-  };
+  // const formatTime = (time: number) => {
+  //   const getSeconds = `0${time % 60}`.slice(-2);
+  //   const minutes = `${Math.floor(time / 60)}`;
+  //   const getMinutes = `0${Number(minutes) % 60}`.slice(-2);
+  //   const getHours = `0${Math.floor(time / 3600)}`.slice(-2);
+  //   return `${getHours}:${getMinutes}:${getSeconds}`;
+  // };
 
   function formatTimeSeconds(seconds: number) {
     const hours = Math.floor(seconds / 3600);
@@ -99,44 +88,6 @@ const DdayList: React.FC<DdayListProps> = ({
   }
 
   // Supabase에서 Challenge 데이터를 가져오는 함수
-  const fetchChallenges = async () => {
-    const {
-      data: { user },
-      error: userError,
-    } = await supabase.auth.getUser();
-    if (userError || !user) {
-      console.error('User not authenticated');
-      return;
-    }
-
-    const { data, error } = await supabase
-      .from('Challenge')
-      .select('*')
-      .eq('User_ID', user.id);
-
-    if (error) {
-      console.error(
-        'Challenge 데이터를 불러오는 데 실패했습니다:',
-        error.message
-      );
-      return;
-    }
-
-    if (data) {
-      console.log(data);
-      const formattedData = data.map((item: any) => ({
-        id: item.Challenge_ID,
-        title: item.Title,
-        dday: calculateDday(item.End_Date),
-        accumulatedTime: item.Accumulated_Time,
-      }));
-      setDdayList(formattedData);
-    }
-  };
-
-  useEffect(() => {
-    fetchChallenges();
-  }, []);
 
   // 디테일 페이지로 이동하는 함수
   const navigateToDetailPage = (challengeId: string) => {
@@ -148,34 +99,37 @@ const DdayList: React.FC<DdayListProps> = ({
       <h1 className="text-2xl font-bold">Challenge</h1>
 
       <div className="my-4">
-        {ddayList.map((item) => (
-          <div
-            key={item.id}
-            onClick={() => onSelectDday(item.id)}
-            className={`flex flex-col items-start justify-between py-2 px-4 rounded-lg mb-2 w-40 ${
-              selectedDdayId === item.id
-                ? 'bg-soft text-black '
-                : 'border border-soft text-black'
-            }`}
-          >
-            <div>
-              <span className="font-bold text-lg">D-{item.dday}</span>
-              <span className="font-normal ml-2">{item.title}</span>
-            </div>
-            <span className="text-sm text-gray-500">
-              누적 시간: {formatTimeSeconds(item.accumulatedTime ?? 0)}
-            </span>
-            <button
-              className="text-xs z-10"
-              onClick={(e) => {
-                e.stopPropagation(); // 부모 div 클릭 이벤트 방지
-                navigateToDetailPage(item.id);
-              }}
+        {ddayList
+          .sort((a, b) => a.dday - b.dday)
+          .map((item) => (
+            <div
+              key={item.id}
+              onClick={() => onSelectDday(item.id)}
+              className={`flex flex-col items-start justify-between py-2 px-4 rounded-lg mb-2 w-40 ${
+                selectedDdayId === item.id
+                  ? 'bg-soft text-black '
+                  : 'border border-soft text-black'
+              }`}
             >
-              수정
-            </button>
-          </div>
-        ))}
+              <div>
+                <span className="font-bold text-lg">D-{item.dday}</span>
+                <span className="font-normal ml-2">{item.title}</span>
+              </div>
+              <span className="text-sm text-gray-500">
+                누적 시간:
+                {formatTimeSeconds(item.accumulatedTime ?? 0)}
+              </span>
+              <button
+                className="text-xs z-10"
+                onClick={(e) => {
+                  e.stopPropagation(); // 부모 div 클릭 이벤트 방지
+                  navigateToDetailPage(item.id);
+                }}
+              >
+                수정
+              </button>
+            </div>
+          ))}
       </div>
 
       <button
